* Move retrieval of base currency for input selector to parent, therefore the request will only be done once for the NftEdit component and not for every input selector component
