+++ admin-front/src/components/Nfts/NftEdit.tsx
@@ -13,12 +13,20 @@ import { NftAside } from './NftAside';
 import { InputSelector } from './Selector/InputSelector';
 import { useStyle } from './useStyle';
 import { FormFieldInputType } from './Selector/types';
+import {
+  getBaseCurrencyFromAPI,
+  getCurrencySymbolDataForCurrency,
+} from '../../shared/hooks/useGetPriceWithCurrency';
+import { CurrencySymbolData } from '../../shared/types/currency';
 
 export const NftEdit = (props: any) => {
+  const notify = useNotify();
   const classes = useStyle();
   const concernedNft = useGetOne('nft', props.id);
 
   const [formFields, setFormFields] = useState<string[]>([]);
+  const [baseCurrencySymbol, setBaseCurrencySymbol] =
+    useState<CurrencySymbolData['symbol']>();
 
   useEffect(() => {
     if (!concernedNft.data) return;
@@ -26,7 +34,13 @@ export const NftEdit = (props: any) => {
     setFormFields(Object.keys(concernedNft.data.allowedActions));
   }, [concernedNft]);
 
-  const notify = useNotify();
+  useEffect(() => {
+    getBaseCurrencyFromAPI({ notify }).then((baseCurrency) => {
+      const currSymbol = getCurrencySymbolDataForCurrency(baseCurrency);
+      setBaseCurrencySymbol(currSymbol?.symbol);
+    });
+  }, [notify]);
+
   const refresh = useRefresh();
   const redirect = useRedirect();
 
@@ -60,22 +74,29 @@ export const NftEdit = (props: any) => {
                 Update an NFT
               </Typography>
               {showFormFields &&
-                formFields.map((fieldName) => (
-                  <Box key={fieldName}>
-                    <InputSelector
-                      attributesName={fieldName}
-                      label={
-                        fieldName[0].toUpperCase() +
-                        fieldName.replace('_', ' ').slice(1)
-                      }
-                      type={
-                        concernedNft.data!.allowedActions[
-                          fieldName
-                        ] as FormFieldInputType
-                      }
-                    />
-                  </Box>
-                ))}
+                formFields.map((fieldName) => {
+                  const label =
+                    fieldName[0].toUpperCase() +
+                    fieldName.replace('_', ' ').slice(1);
+                  const type = concernedNft.data!.allowedActions[
+                    fieldName
+                  ] as FormFieldInputType;
+
+                  const useBaseCurrencySymbol =
+                    type === 'number' && label === 'Price';
+                  return (
+                    <Box key={fieldName}>
+                      <InputSelector
+                        attributesName={fieldName}
+                        label={label}
+                        type={type}
+                        baseCurrencySymbol={
+                          useBaseCurrencySymbol ? baseCurrencySymbol : undefined
+                        }
+                      />
+                    </Box>
+                  );
+                })}
             </Box>
           </Box>
         </SimpleForm>
