+++ admin-front/src/components/Nfts/Selector/InputSelector.tsx
@@ -1,4 +1,4 @@
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import {
   BooleanInput,
   DateTimeInput,
@@ -10,28 +10,11 @@ import {
   RadioButtonGroupInput,
   SelectArrayInput,
   TextInput,
-  useNotify,
 } from 'react-admin';
 import { useStyle } from '../useStyle';
 import { InputSelectorProps } from './types';
-import {
-  getBaseCurrencyFromAPI,
-  getCurrencySymbolDataForCurrency,
-} from 'shared/hooks/useGetPriceWithCurrency';
-import { CurrencySymbolData } from 'shared/types/currency';
 
 export const InputSelector: FC<InputSelectorProps> = ({ ...props }) => {
-  const notify = useNotify();
-  const [currencySymbol, setCurrencySymbol] =
-    useState<CurrencySymbolData['symbol']>();
-
-  useEffect(() => {
-    getBaseCurrencyFromAPI({ notify }).then((baseCurrency) => {
-      const currSymbol = getCurrencySymbolDataForCurrency(baseCurrency);
-      setCurrencySymbol(currSymbol?.symbol);
-    });
-  }, [notify]);
-
   const validateNumber = [number('expecting a number'), minValue(0)];
   const validateDate = (value: any) => {
     if (value < new Date().getTime()) return 'Date must be in the future';
@@ -39,8 +22,6 @@ export const InputSelector: FC<InputSelectorProps> = ({ ...props }) => {
   };
 
   const classes = useStyle();
-
-  const showCurrency = props.type === 'number' && props.label === 'Price';
 
   const { type } = props;
 
@@ -60,18 +41,19 @@ export const InputSelector: FC<InputSelectorProps> = ({ ...props }) => {
         multiline
       />
     );
-  if (type === 'number')
+  if (type === 'number') {
+    const label = props.baseCurrencySymbol
+      ? props.label + ` in ${props.baseCurrencySymbol}`
+      : props.label;
     return (
       <TextInput
         source={`attributes.${props.attributesName}`}
-        label={
-          showCurrency && currencySymbol
-            ? props.label + ` in ${currencySymbol}`
-            : props.label
-        }
+        label={label}
         validate={validateNumber}
       />
     );
+  }
+
   if (type === 'boolean')
     return (
       <BooleanInput
